feat: the five migrated recipes, and the pages they replace (#133)
Slice 2 of 6. Five pages become five recipes and are deleted; the named routes
that served them go with them.

`domain_type_page` splits in two, because it always was two things: a `label`
extractor over a non-`String` type, and a dropdown assembled from the raw
controller. The first becomes `domain_type_recipe.dart`, and it gains the
multi-select half it never had — both cardinalities take the same `label`
callback and differ only in what they hold, which is the one way "cardinality is
a type, not a flag" is visible rather than asserted.

The second becomes `build_your_own_recipe.dart`, and it now uses
`TextItemPresentation` rather than hand-drawn rows. That is the pair the README
advertises, and drawing the rows by hand was quietly giving up what the
presentation exists for: `buildItem`'s second argument is what a screen reader
hears, and on screen the difference is a colour.

`bare_anchor_recipe.dart` gains a third variant, the default constructor. Not
symmetry for its own sake — `anchorBuilder` asserts against `width`, `minWidth`,
`maxWidth`, `expand` and `trailing`, so a playground that owns those knobs can
never put that combination on screen. It has to live in a recipe or nowhere.

`multi_select_recipe.dart` is the first recipe with knobs, and it is where the
split decided for this map meets real code: `searchable` and *Drop Solaris*
live outside in a notifier, because the stage and the knob pane are two views of
one value drawn on the same frame. The ticked `Set` stays inside the recipe,
because a caller holding the set is what this package's API *is*, and a file
taking its selection from an injected object is not pasteable.

`overlay_lifetime_recipe.dart` absorbs `bug_test_page` and `result_page`, on its
own nested `Navigator`. Roughly two hundred lines of instruction card do not
come with it: the Code pane reads the file's real bytes, header comment
included, so the narrative is a dartdoc now — and unlike a widget, it travels
with the paste.

The playground's four `AppBar` nav actions are gone. They pointed at routes that
no longer exist, and `flutter analyze` had nothing to say about it: they are
string literals, so the failure would have been a runtime throw in the one
surface that still worked.

API coverage rises 132/203 -> 136/203 (67.0%), and the floor rises with it. That
number is the ratchet doing its job: five pages died in this commit, and had the
recipes carried less than they did, it would have gone red before review noticed.

Docs move where their subject moved. `documentation/api_reference.md` and
`documentation/migration.md` cited `domain_type_page` as the working example of
`DropdownOverlayController`, which is now the Build Your Own recipe. The README's
Build Your Own bullet named one of three presentations. `CHANGELOG.md` still
names the page and is left alone — a published entry is superseded, never
rewritten, and editing it would forge the past.

Closes #133
