+++ example/lib/app/destinations.dart
@@ -4,29 +4,29 @@ library;
 import 'package:flutter/widgets.dart';
 import 'package:flutter_example_template/flutter_example_template.dart';
 
-import '../pages/bare_anchor_page.dart';
-import '../pages/bug_test_page.dart';
-import '../pages/domain_type_page.dart';
-import '../pages/multi_select_page.dart';
 import '../pages/playground_page.dart';
+import '../recipes/bare_anchor_recipe.dart';
 import '../recipes/basic_recipe.dart';
+import '../recipes/build_your_own_recipe.dart';
+import '../recipes/domain_type_recipe.dart';
+import '../recipes/overlay_lifetime_recipe.dart';
+import 'recipe_knobs.dart';
 
 /// The destinations of *this* app, and whatever state sits behind them.
 ///
 /// The split is the point: [ShellPage] draws a menu, a stage and a knob region
 /// and knows nothing about dropdowns, and this file knows nothing about layout.
 ///
-/// **Every entry here is a [RouteDestination], and every one of them is
-/// temporary.** They are the pages this example had before the shell arrived —
-/// each a full page with its own `Scaffold`, which is exactly what
-/// [ShellCategory.pages] means: a hosting kind rather than a subject. They are
-/// listed rather than deleted so this slice can land with no recipe existing,
-/// and they are replaced one at a time as the recipes arrive.
-///
-/// Nothing here carries a `source:` yet. The Code pane is the affordance of the
-/// pasteable claim, and none of these pages makes it.
+/// **What lives here and what does not.** A recipe's *knobs* live here, because
+/// the stage and the knob pane are two views of one value drawn on the same
+/// frame. A recipe's *selection* does not: it lives in the recipe's own
+/// `State`, because a file that took its selection from an injected object
+/// would not be pasteable — and because a caller holding the selection is
+/// exactly what this package's API is.
 class DropdownDestinations implements ShellDestinations {
   DropdownDestinations();
+
+  final _multiSelectKnobs = MultiSelectKnobs();
 
   /// Every destination, in menu order.
   ///
@@ -49,48 +49,64 @@ class DropdownDestinations implements ShellDestinations {
       // not there.
       knobs: (context) => const SizedBox.shrink(),
     ),
+    StageDestination(
+      id: 'domain-type',
+      label: 'Domain type',
+      category: ShellCategory.recipes,
+      source: 'lib/recipes/domain_type_recipe.dart',
+      stage: (context) => const DomainTypeRecipe(),
+      knobs: (context) => const SizedBox.shrink(),
+    ),
+    StageDestination(
+      id: 'multi-select',
+      label: 'Multi-select',
+      category: ShellCategory.recipes,
+      source: 'lib/recipes/multi_select_recipe.dart',
+      stage: (context) => MultiSelectStage(knobs: _multiSelectKnobs),
+      knobs: (context) => MultiSelectKnobPane(knobs: _multiSelectKnobs),
+    ),
+    StageDestination(
+      id: 'bare-anchor',
+      label: 'Bare anchor',
+      category: ShellCategory.recipes,
+      source: 'lib/recipes/bare_anchor_recipe.dart',
+      stage: (context) => const BareAnchorRecipe(),
+      knobs: (context) => const SizedBox.shrink(),
+    ),
+    StageDestination(
+      id: 'build-your-own',
+      label: 'Build your own',
+      category: ShellCategory.recipes,
+      source: 'lib/recipes/build_your_own_recipe.dart',
+      stage: (context) => const BuildYourOwnRecipe(),
+      knobs: (context) => const SizedBox.shrink(),
+    ),
+    StageDestination(
+      id: 'overlay-lifetime',
+      label: 'Overlay lifetime',
+      category: ShellCategory.recipes,
+      source: 'lib/recipes/overlay_lifetime_recipe.dart',
+      stage: (context) => const OverlayLifetimeRecipe(),
+      knobs: (context) => const SizedBox.shrink(),
+    ),
     RouteDestination(
       id: 'playground',
       label: 'Every setting',
       category: ShellCategory.pages,
       open: (context) => const PlaygroundPage(),
     ),
-    RouteDestination(
-      id: 'multi-select',
-      label: 'Multi-select',
-      category: ShellCategory.pages,
-      open: (context) => const MultiSelectPage(),
-    ),
-    RouteDestination(
-      id: 'domain-type',
-      label: 'Domain type',
-      category: ShellCategory.pages,
-      open: (context) => const DomainTypePage(),
-    ),
-    RouteDestination(
-      id: 'bare-anchor',
-      label: 'Bare anchor',
-      category: ShellCategory.pages,
-      open: (context) => const BareAnchorPage(),
-    ),
-    RouteDestination(
-      id: 'overlay-teardown',
-      label: 'Overlay teardown',
-      category: ShellCategory.pages,
-      open: (context) => const DropdownBugTestPage(),
-    ),
   ];
 
   /// Releases whatever the destinations hold.
   ///
-  /// **Empty, and only for as long as every destination is a route.** A
-  /// [RouteDestination] is pointed at rather than absorbed, so its page owns
-  /// its own `State` and Flutter disposes it. The first `StageDestination` to
-  /// arrive changes that: a recipe's knob notifier, and the
-  /// `DropdownOverlayController` the Build Your Own recipe holds, are both
-  /// owned here and both have to be released here. A leaked `OverlayEntry` is
-  /// the consumer's whole app — a dead layer swallowing taps — and it throws
-  /// nothing and fails no test.
+  /// The knob notifier, and nothing else. Every recipe's own state is in its
+  /// own `State`, which Flutter disposes — including the
+  /// `DropdownOverlayController` the Build Your Own recipe holds, which is
+  /// disposed in that recipe rather than here because the recipe has to be
+  /// pasteable with its lifetime intact. A leaked `OverlayEntry` is the
+  /// consumer's whole app, and it throws nothing and fails no test.
   @override
-  void dispose() {}
+  void dispose() {
+    _multiSelectKnobs.dispose();
+  }
 }
