+++ example/lib/pages/playground_page.dart
@@ -400,28 +400,6 @@ class _PlaygroundPageState extends State<PlaygroundPage> {
       appBar: AppBar(
         backgroundColor: Theme.of(context).colorScheme.inversePrimary,
         title: const Text('Dropdown Playground'),
-        actions: [
-          IconButton(
-            icon: const Icon(Icons.checklist),
-            tooltip: 'Multi-Select Filter',
-            onPressed: () => Navigator.pushNamed(context, '/multi-select'),
-          ),
-          IconButton(
-            icon: const Icon(Icons.travel_explore),
-            tooltip: 'Domain Types & Controller',
-            onPressed: () => Navigator.pushNamed(context, '/domain-type'),
-          ),
-          IconButton(
-            icon: const Icon(Icons.manage_search),
-            tooltip: 'Bare Anchor (inline in a field)',
-            onPressed: () => Navigator.pushNamed(context, '/bare-anchor'),
-          ),
-          IconButton(
-            icon: const Icon(Icons.bug_report),
-            tooltip: 'Overlay Bug Test',
-            onPressed: () => Navigator.pushNamed(context, '/dropdown-bug-test'),
-          ),
-        ],
       ),
       body: isWide
           ? Row(
